+++ Tests/Runtime/DucktionSingletonTest.cs
@@ -1,8 +1,6 @@
 ﻿using NUnit.Framework;
 using TheRealIronDuck.Ducktion.Editor.Tests.Editor;
 using TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs;
-using TheRealIronDuck.Ducktion.Exceptions;
-using UnityEngine;
 
 namespace TheRealIronDuck.Ducktion.Tests
 {